moved UIColor extension to its own file
